+++ lib/nanoc-webpack.rb
@@ -1,1 +1,3 @@
+# frozen_string_literal: true
+
 require_relative "nanoc/webpack"
