+++ lib/nanoc/webpack.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require "nanoc"
 module Nanoc::Webpack
   require_relative "webpack/filter"
